Moving stop to test service
